(#3324) Ported yumrepo unit tests to spec
* Port test from test/ral/type/yumrepo.rb into yumrepo_spec.rb
* Delete old test and supporting files
